LLVM HelloWorld - Initial commit
LLVM HelloWorld, the project 'llvm-first-step' is the demo project
described in chapter 3 of 'Getting Started with LLVM Core Libraries'.
The project was adopted for LLVM 7.0

CMake 3.11 is used as a build tool. A path to a debug build of LLVM 7.0
must be specified as a value of the 'LLVM_DEBUG_PREFIX' variable.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>