Add LLVM frontend demo - clang based diagnostic
Clang (LLVM C/C++ Frontend) HelloWorld, the project 'llvm-front-clang-diagnostic'
is the demo project described in chapter 4 of 'Getting Started with LLVM
Core Libraries'. The project was adopted for LLVM 7.0 and uses a
configuration header generated by CMake.

CMake 3.11 is used as a build tool. A path to a build of LLVM 7.0
must be specified as a value of the 'LLVM_INSTAL_PREFIX' variable.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>