Add a simple Clang based frontend
Clang (LLVM C/C++ Frontend) simple frontend, the project
'llvm-front-clang-simple-frontend' is the demo project described in
chapter 4 of 'Getting Started with LLVM Core Libraries'. The project
was adopted for LLVM 7.0.

CMake 3.11 is used as a build tool. A path to a build of LLVM 7.0
must be specified as a value of the 'LLVM_INSTAL_PREFIX' variable.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>