Add a simple LLVM IR generator
Simple LLVM IR generator for a 'get minimum' function. The code is based
on the one from chapter 5 of 'Getting Started with LLVM Core Libraries'.
The project was adopted for LLVM 7.0 and the following changes has been
made:

 * The code generates the 'min' function, not the 'sum' one.
 * Some metadata ('llvm.ident') is added to the IR module
 * We set 'DSO Local' as true
 * 4 function attributes are set: 'noinline', 'nounwind' and 'uwtable'.

Note: an output file must be flushed otherwise the IR module would be
corrupted.

CMake 3.11 is used as a build tool. A path to a build of LLVM 7.0
must be specified as a value of the 'LLVM_INSTAL_PREFIX' variable.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>