Add a custom IR pass
A custom IR pass ('FunctionArgumentCount') demonstrates how to develop a
custom pass and build it as a part of the LLVM source tree.

There is no dynamic linking on Windows (this is OS weirdness), so we
cannot use the plugins at all, unfortunately. The solution is just to add
a pass to the LLVM source tree and built them together. For detailed
instruction how to register a new created pass into the
'llvm/InitializePasses.h' header and the 'opt' tool see the README.md file.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>