Ignore the CLion project folder
CLion is a cross-platform C/C++ IDE from JetBrains (they are famous in
the Java community as the authors of IntelliJ IDEA).
The configuration of a CLion project is stored in an .idea folder
inside the root of the project. The configuration is developer-specific,
so the folder must be ignored by Git.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>