Add a machine pass to count emitted instructions
A machine function pass ('CountInstructionsPass') demonstrates how
to develop a machine pass and build it as a part of the LLVM source tree.

There is no dynamic linking on Windows, so we cannot use plugins at all,
unfortunately. The solution is just to add a pass to the LLVM source
tree and build them together. For detailed instruction how to register
the new created pass into the backend see the README.md file.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>