Add a patch to easy register passes in LLVM tree
To build a machine pass as a part of the LLVM source tree a number of
manual task must be done and target-related header files changed. If a
LLVM Git repo is used (from the official website or GitHub unofficial
mirror), the 'git apply' command is available and can be used to do this
stuff. The corresponding patch is published into the 'patches' directory
of the 'llvm-backend-x86-machine-passes' experiment.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>