Optimize using of LLVM_DEBUG macro
The FunctionArgumentUsagePass analysis pass is modified to show how to
use the 'LLVM_DEBUG' macro: the macro can wrap not only a single line of
code but also a whole block.

'dumFunctionsArg' is a debug only function, the function is re-designed
as 'LLVM_DEBUG' free thus a call to the function must be wrapped into
'LLVM_DEBUG'.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>