Add GTest-based unit tests
GTest will be loaded into the CMake context as described in the
following amazing blog post: Building GoogleTest and GoogleMock directly
in a CMake project https://crascit.com/2015/07/25/cmake-gtest/

The module 'DownloadProject' is used to download the latest master of
CMake and incorporate it into the build three (see 'DownloadProject.cmake'
and 'DownloadProjectCmakeLists.cmake.in'). An example
('simple-generator-test.cpp') demonstrates how to develop a unit test
with mocks for LLVM (in our case, for an IR-code generator). The test can
work with enabled assertions as well as with disabled ones. To enable
assertions, a piece of code has been added to 'CMakeLists.txt' (starting
from 'if(NOT uppercase_CMAKE_BUILD_TYPE STREQUAL "DEBUG")', the code was
borrowed from LLVM itself (see 'cmake/modules/HandleLLVMOptions.cmake').

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>