Introduce the 'avg-unique-instrs' command line arg
The command line argument 'avg-unique-instrs' is used to let the pass
create a default-sized map to store numbers of each instruction in a
function under analysis. In order to get the balance between memory
usage and numbers of memory allocation requests, the parameter should be
set to a value close to the average number of unique instructions in a
function under analysis.

Yes, I understand the functions can be different, so the user can make
his/her own decision in favour to time or memory.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>